pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = "User microservice of black-widow project. Built with AWS Chalice"
authors = ["jakub-figat <jakub.figat@neontri.com>"]

[tool.poetry.dependencies]
python = "^3.9"
chalice = "^1.27.0"
SQLAlchemy = "^1.4.36"
pydantic = {extras = ["email"], version = "^1.9.0"}
boto3 = "^1.22.9"
mypy-boto3-dynamodb = "^1.22.8"
passlib = "^1.7.4"
bcrypt = "^3.2.2"
PyJWT = "^2.3.0"
mypy-boto3-s3 = "^1.22.8"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
coverage = "^6.3.2"
pytest = "^7.1.1"
prospector = "^1.7.7"
pytest-mock = "^3.7.0"


[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | env
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
line_length = 119
lines_after_imports = 2
profile = "black"
skip_glob = ["env/*"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
